Moves DEBUG code paths from the main `SceneDelegate` to new `DebuggingSceneDelegate` subclass to separate debug and test specific code from production code.
